requirements_tests.txt
-r requirements.txt
pillow>=9.1.0
pytest>=6.0.0
pytest-benchmark
pytest-xdist
pytest-lazy-fixtures
pyopengl
glfw
PyOpenAL
filelock
cryptography
GitPython
